refactor: stories import 경로 수정
