+++ src/stories/components/ChipContainer.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta } from '@storybook/react';
 
-import ChipContainer from '@/src/app/main/_component/ChipContainer';
+import ChipContainer from '@/src/app/(main)/_component/ChipContainer';
 
 const meta: Meta<typeof ChipContainer> = {
   title: 'components/ChipSelect',
