+++ src/stories/components/Popup.stories.tsx
@@ -1,5 +1,3 @@
-import { Dispatch, SetStateAction } from 'react';
-
 import { Meta } from '@storybook/react';
 
 import Popup from '@/src/app/_components/Popup';
@@ -12,8 +10,6 @@ const meta: Meta<typeof Popup> = {
 
 export default meta;
 
-export const DefaultTemplate = (
-  setSee: Dispatch<SetStateAction<string | null>>,
-) => {
-  return <Popup setSee={setSee} />;
+export const DefaultTemplate = () => {
+  return <Popup />;
 };
