+++ src/stories/components/PostItem.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import PostItem from '@/src/app/main/_component/PostItem';
+import PostItem from '@/src/app/(main)/_component/PostItem';
 
 import { PageLayoutDecorator } from '../decorator';
